+++ resources/js/humhub.wiki.Page.js
@@ -13,5 +13,9 @@ humhub.module('wiki.Page', function(module, require, $) {
         });
     };
 
+    Page.print = function() {
+        window.print();
+    }
+
     module.export = Page;
 });