+++ config.php
@@ -7,6 +7,9 @@ return [
     'id' => 'wiki',
     'class' => 'humhub\modules\wiki\Module',
     'namespace' => 'humhub\modules\wiki',
+    'urlManagerRules' => [
+        ['class' => 'humhub\modules\wiki\components\WikiPageUrlRule'],
+    ],
     'events' => array(
         ['class' => Menu::class, 'event' => Menu::EVENT_INIT, 'callback' => ['humhub\modules\wiki\Events', 'onSpaceMenuInit']],
         ['class' => ProfileMenu::class, 'event' => ProfileMenu::EVENT_INIT, 'callback' => ['humhub\modules\wiki\Events', 'onProfileMenuInit']],