+++ docs/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 - Enh #148: Store folding state for categories per user
 - Enh #180: Option to add wiki pages to Space menu
 - Enh #62: Add print version
-
+- Enh #22: Diff view for Wiki page changes
 
 1.5.3  (April 16, 2021)
 ----------------------